modal tracker will be deleted also when user refresh the page

+++ src/main/java/com/forum/application/service/ModalTrackerService.java
@@ -37,6 +37,7 @@ public class ModalTrackerService {
     
     void deleteTrackerOfUserById(int userId, ModalTracker.Type type) {
         ModalTracker modalTracker = getTrackerOfUserById(userId);
+        if (modalTracker == null) return;
         if (modalTracker.getType() == type) {
             modalTrackerRepository.deleteById(userId);
             log.debug("Deleting modal tracker for receiver with id of {} success!", userId);
